Commencement initiation combat

+++ src/main/java/org/example/Adversaire.java
@@ -4,7 +4,7 @@ public class Adversaire extends Personnage {
 
 
 
-    public Adversaire(String nom, char sexe, int age, int taille, char classe , int force , int chance , int agilite) {
+    public Adversaire(String nom, char sexe, int age, int taille, char classe , float force , float chance , float agilite) {
         super(nom, sexe, age, taille, classe);
